refactor(ui): split CredentialFormPage into reusable form building blocks
Phase 1 of the credentials revamp v3 (sheet-edit + toolkit-anchored
add). Pure refactor — no behaviour change, no DOM/text change. The
existing /credentials/new and /credentials/:id/edit routes work
identically; all 13 CredentialFormPage tests pass without edits.

The 1267-line CredentialFormPage was a route, page chrome, deeplink
parser, query orchestrator, picker, form, mutation handler, and
sticky footer all in one file. Split into:

- lib/credentials/schemes.ts — pure helpers for OpenAPI
  securitySchemes parsing (parseSchemeOptions, isCompoundApiKey,
  schemeTypeFromRaw, etc). No React, easy to unit test, shared by
  every credential surface.
- hooks/useApiSchemes.ts — resolves schemes for a selected API
  (local detail vs catalog spec fetch via spec_url, with 5-min
  staleTime). Returns localDetail and spec separately so the next
  hook can reuse them.
- hooks/useApiServerVarDefs.ts — normalises server-variable defs
  across local + catalog sources.
- components/credentials/form/ApiPicker.tsx — debounced search,
  local/catalog grouping, autofocus.
- components/credentials/form/SchemePillBar.tsx — pure presentation,
  hidden when only one auth method.
- components/credentials/form/ServerVariablesFields.tsx — Input or
  Select per def, parent owns values map.
- components/credentials/form/OAuthBrokerFields.tsx — owns the
  connect-link mutation and the broker-loading/has-broker/no-broker
  branches.
- components/credentials/form/AdvancedBrokerFields.tsx — collapsible
  scheme JSON + routes textarea pair, parent owns open state.
- components/credentials/form/AuthTypeFields.tsx — branches across
  basic / bearer / apiKey / unknown / compound, no internal state.
- components/credentials/form/CredentialFormFields.tsx — the
  orchestrator. Owns local form state, the create + update
  mutations, and the cross-tab event emission. Host-agnostic:
  takes onBack and onSaved callbacks instead of calling navigate
  directly. This is the seam the upcoming sheet (Phase 2) and
  toolkit-anchored dialog (Phase 3) compose on.

CredentialFormPage shrinks to a thin route shell: URL parsing,
existing/existingApi/paramApi queries, page chrome, step indicator.
All form behaviour lives in CredentialFormFields.

The only deliberate API change is in CredentialFormFields.onSaved —
it now receives {id, api_id} so non-page hosts (sheet, dialog) can
chain follow-up calls (e.g. toolkit binding). The page host ignores
the payload and navigates as before.
