+++ ui/src/components/credentials/index.ts
@@ -4,3 +4,16 @@ export { PipedreamCard } from './PipedreamCard';
 export { StatusDot } from './StatusDot';
 export { TestConnectionButton } from './TestConnectionButton';
 export type { CredentialStatus } from './StatusDot';
+
+// Form building blocks. Composed by `CredentialFormPage` today and by
+// the upcoming sheet-based edit + toolkit-anchored add surfaces.
+export {
+	ApiPicker,
+	CredentialFormFields,
+	SchemePillBar,
+	ServerVariablesFields,
+	OAuthBrokerFields,
+	AdvancedBrokerFields,
+	AuthTypeFields,
+} from './form';
+export type { CredentialFormPrefill, CredentialFormFieldsProps } from './form';
